Changes:
Added a main web page with 20 pictures and 20 checkboxes for my website.
Added 4 proper links on each web page.
Added a web page that creates a new user account.
Added a web page that will login to a user account.
Added a web page that functions as a wishlist for a user.
Added a web page that is meant to be for an administrator that can
search for users and have it show all usernames, user IDs, and how much
money the users have in their account.
Added a web page that functions as a shopping cart for when the user has
selected items from the main page.

Swapped out mysql-connector-java-8.0.12 for sql-connector-java-8.0.17.
Renamed the table PersonalProjectBram to UserTable.

In Progress:
Main web page that will add the items a user selects to a table called
ItemsTable.
Login web page along with logout functionality.
Wishlist web page.
Shopping cart web page.